win_template - use shared doc fragment for shared options
