+++ test/integration/targets/win_template/templates/foo.utf-8.j2
@@ -1,0 +1,1 @@
+“café” - {{ templated_var }}
